requirements.txt
absl-py==1.4.0
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
array-record==0.4.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==46.0.1
cycler==0.12.1
dacite==1.8.1
dask==2023.3.1
dataclass_array==1.4.1
decorator==5.2.1
descartes==1.1.0
dm-tree==0.1.8
einops==0.8.1
einsum==0.3.0
etils==1.3.0
filelock==3.14.0
fire==0.7.1
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.70.0
h5py==3.11.0
idna==3.10
imageio==2.35.1
immutabledict==2.2.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
jmespath==0.10.0
joblib==1.4.2
keras==2.12.0
kiwisolver==1.4.7
lark==1.3.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.31.0
locket==1.0.0
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.1
mdurl==0.1.2
ml-dtypes==0.2.0
mmcv==1.4.0
mmcv-full==1.4.0
mmdet==2.20.0
mmdet3d==0.0.0
model-index==0.1.11
mpi4py==3.0.3
msgpack==1.1.1
msgpack-numpy==0.4.8
multimethod==1.10
networkx==3.1
numba==0.48.0
numpy==1.23.0
nuscenes-devkit==1.1.11
oauthlib==3.3.1
opencv-python==4.12.0.88
opendatalab==0.0.10
OpenEXR==1.3.9
openmim==0.3.9
openxlab==0.1.2
opt_einsum==3.4.0
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==1.5.3
partd==1.4.1
Pillow==9.2.0
platformdirs==4.3.6
plotly==5.13.1
promise==2.3
protobuf==3.20.3
psutil==7.1.0
pyarrow==10.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.7
pycparser==2.23
pycryptodome==3.23.0
Pygments==2.19.2
pyparsing==3.1.4
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==27.1.0
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.4.2
rsa==4.9.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
Shapely==1.8.5.post1
six==1.17.0
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-addons==0.21.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
tensorflow-probability==0.19.0
tensorpack==0.11
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==1.10.1+cu113
torchaudio==0.10.1+cu113
torchpack==0.3.1
torchvision==0.11.2+cu113
tqdm==4.65.2
trimesh==4.8.2
typeguard==2.13.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
visu3d==1.5.1
waymo-open-dataset-tf-2-12-0==1.6.5
Werkzeug==3.0.6
wrapt==1.14.2
yapf==0.32.0
zipp==3.20.2
